Separate channel nodes
